Data frame of all flow data used in anlz_nps_gaged and anlz_nps_ungaged

Data frame of all flow data used in anlz_nps_gaged and anlz_nps_ungaged

Usage

allflo

Format

A data.frame of monthly mean daily flow data for select basins

Details

Monthly flow data at select stations used for estimating non-point source gaged and ungaged loads. Created using the util_nps_getflow function. Includes data from the USGS API using util_nps_getusgsflow and from external sources using util_nps_getextflow. The data frame contains the following columns:

basin: Character string for the basin or gauge location

yr: Year of the observation

mo: Month of the observation

flow_cfs: Numeric value for the average daily flow in cubic feet per second (cfs)

See also

Examples

allflo
#> # A tibble: 612 × 4
#>    basin       yr    mo flow_cfs
#>    <chr>    <dbl> <dbl>    <dbl>
#>  1 02299950  2021     1     18.7
#>  2 02299950  2021     2     41.9
#>  3 02299950  2021     3     17.1
#>  4 02299950  2021     4     73.7
#>  5 02299950  2021     5      6.8
#>  6 02299950  2021     6     18.7
#>  7 02299950  2021     7    362. 
#>  8 02299950  2021     8    132. 
#>  9 02299950  2021     9    317. 
#> 10 02299950  2021    10     64.7
#> # ℹ 602 more rows