Domestic and industrial point source facilities, including industrial with material losses

Domestic and industrial point source facilities, including industrial with material losses

Usage

facilities

Format

A data.frame

Examples

facilities
#> # A tibble: 101 × 11
#>    bayseg basin  entity facname source entityshr facnameshr permit facid coastco
#>     <dbl> <chr>  <chr>  <chr>   <chr>  <chr>     <chr>      <chr>  <chr> <chr>  
#>  1      2 02303… Alpha… Alpha/… PS - … aoc       aoc        FL002… FL00… 327    
#>  2      7 202-7  Brade… City o… PS - … bradenton bradenton  FL002… MC757 736    
#>  3      7 202-7  Brade… City o… PS - … bradenton bradenton  FL002… MC757 736    
#>  4      2 02301… Brews… Brewst… PS - … brewster  brewster   FL013… FL01… 549a   
#>  5      2 02304… Busch… Busch … PS - … busch     busch      FL018… FL01… 191a   
#>  6      2 206-2  CSX    CSX - … PS - … csx       rocknewp   FL045… FL04… 504    
#>  7      2 02303… CSX    CSX Wi… PS - … csx       winston    FL003… 4053… 423    
#>  8      2 NA     CSX    Newport PS - … csx       newp       NA     NA    NA     
#>  9      2 NA     CSX    Rockpo… PS - … csx       rock       NA     NA    NA     
#> 10      1 206-1  Clear… City o… PS - … clearwat… east       FL002… PC691 480    
#> # ℹ 91 more rows
#> # ℹ 1 more variable: coastid <chr>