Data frame of distances of drainage basin locations to National Weather Service (NWS) sites

Data frame of distances of drainage basin locations to National Weather Service (NWS) sites

Usage

nps_distance

Format

A data.frame

Details

Used for estimating non-point source (NPS) ungaged loads. The data frame contains the following columns:

target: Numeric identifier for the drainage basin

targ_x: Numeric value for the x-coordinate of the drainage basin location (WGS 84, UTM Zone 17N, CRS 32617)

targ_y: Numeric value for the y-coordinate of the drainage basin location (WGS 84, UTM Zone 17N, CRS 32617)

matchsit: Numeric for the NWS site that matches the drainage basin location

distance: Numeric value for the distance (m) between the drainage basin coordinate and NWS site

invdist2: Numeric value for the inverse distance squared (1/m^2) between the drainage basin coordinate and NWS site

Examples

nps_distance
#>       target targ_x  targ_y matchsit         distance             invdist2
#> 1   02299950 386840 3047697      228 47117.0193454552 4.50447708806558e-10
#> 2   02299950 386840 3047697      478  49548.931492011 4.07315946391767e-10
#> 3   02299950 386840 3047697      940 43814.4066261315 5.20914121687418e-10
#> 4   02299950 386840 3047697      945 33546.9593256975 8.88574151975041e-10
#> 5   02299950 386840 3047697     3153 2246.60855513372 1.98127693298331e-07
#> 6   02299950 386840 3047697     6065 37257.6405586827 7.20392699315623e-10
#> 7   02299950 386840 3047697     6880 26968.4498998367 1.37495356266211e-09
#> 8   02299950 386840 3047697     9401 32580.3603724698  9.4208097294867e-10
#> 9   02300500 379928 3060372      478 46013.0708060221 4.72321335640097e-10
#> 10  02300500 379928 3060372      940 41289.9632477434 5.86558060671061e-10
#> 11  02300500 379928 3060372      945 34154.1217571174  8.5726235976074e-10
#> 12  02300500 379928 3060372     3153 13577.1154889395 5.42481554203143e-09
#> 13  02300500 379928 3060372     6065 46870.2447294656  4.5520346335947e-10
#> 14  02300500 379928 3060372     6880 21588.9498818261 2.14554171975086e-09
#> 15  02300500 379928 3060372     7205 39970.4181864538 6.25925458174149e-10
#> 16  02300500 379928 3060372     7886 43101.6523233159 5.38284855525742e-10
#> 17  02300500 379928 3060372     8788 45777.6426872333 4.77191995388684e-10
#> 18  02300500 379928 3060372     9401 40977.7096114461 5.95531364303964e-10
#> 19  02300500 379928 3060372     2806 43101.6523233159 5.38284855525742e-10
#> 20  02300530 366067 3067562      520  40093.184308558 6.22098134602164e-10
#> 21  02300530 366067 3067562      940 34961.1669427667 8.18141004399673e-10
#> 22  02300530 366067 3067562      945  32517.972415266 9.45699334396949e-10
#> 23  02300530 366067 3067562     1632 46786.8497443459  4.5682765984071e-10
#> 24  02300530 366067 3067562     3153  28476.991203426 1.23313833157873e-09
#> 25  02300530 366067 3067562     3986 48430.3510208216 4.26348526533097e-10
#> 26  02300530 366067 3067562     6880 16944.3635466193 3.48296795234809e-09
#> 27  02300530 366067 3067562     7205 38988.8449816098 6.57838460060212e-10
#> 28  02300530 366067 3067562     7886 27776.4191356625 1.29612678691492e-09
#> 29  02300530 366067 3067562     8788 31612.0747974567 1.00067718526956e-09
#> 30  02300530 366067 3067562     2806 27776.4191356625 1.29612678691492e-09
#> 31  02300700 370592 3071162      478 49732.5474613959 4.04313821956778e-10
#> 32  02300700 370592 3071162      520 38612.6177952234 6.70720387560913e-10
#> 33  02300700 370592 3071162      940 40644.7065311093 6.05329745966853e-10
#> 34  02300700 370592 3071162      945 37560.6456946629 7.08816634491767e-10
#> 35  02300700 370592 3071162     1632 48720.7356676806 4.21281439018397e-10
#> 36  02300700 370592 3071162     3153 27830.4447143771 1.29109948434997e-09
#> 37  02300700 370592 3071162     3986 43892.8584737882 5.19053675353545e-10
#> 38  02300700 370592 3071162     6880  22097.666867794 2.04789250143702e-09
#> 39  02300700 370592 3071162     7205 33475.8787487349 8.92351643179617e-10
#> 40  02300700 370592 3071162     7886 31657.6260164909 9.97799567580608e-10
#> 41  02300700 370592 3071162     8788 31514.9123114757 1.00685700024247e-09
#> 42  02300700 370592 3071162     2806 31657.6260164909 9.97799567580608e-10
#> 43  02301000 400160 3088462      478 17964.8188691119 3.09852007647507e-09
#> 44  02301000 400160 3088462     3153 40738.1443490005 6.02556137737597e-10
#> 45  02301000 400160 3088462     3986  33432.845661116 8.94650302440248e-10
#> 46  02301000 400160 3088462     4707 33119.4261574684 9.11663134301739e-10
#> 47  02301000 400160 3088462     5973  40830.634601975 5.99829388607544e-10
#> 48  02301000 400160 3088462     7205 15926.3947583877 3.94243959344774e-09
#> 49  02301000 400160 3088462     9401  43626.433924858 5.25412705414557e-10
#> 50  02301300 391299 3070102      478 31183.4111347684 1.02837795600688e-09
#> 51  02301300 391299 3070102      945 49108.2052003532 4.14659744946158e-10
#> 52  02301300 391299 3070102     3153 20802.4941773817 2.31083630358218e-09
#> 53  02301300 391299 3070102     3986 45868.4554350809 4.75304326951382e-10
#> 54  02301300 391299 3070102     6880 36175.2014507176 7.64149060417598e-10
#> 55  02301300 391299 3070102     7205 29419.0116931212 1.15543053930637e-09
#> 56  02301300 391299 3070102     8788 48568.6307095434 4.23924269771655e-10
#> 57  02301300 391299 3070102     9401 34948.6071253205 8.18729156486111e-10
#> 58  02301500 393913 3081151      478 24603.5780527955 1.65197495679313e-09
#> 59  02301500 393913 3081151      520 47869.7300284846 4.36393269275669e-10
#> 60  02301500 393913 3081151     3153 32133.6721368723 9.68454638434555e-10
#> 61  02301500 393913 3081151     3986 36356.8229772625 7.56533473957977e-10
#> 62  02301500 393913 3081151     4707 42426.7009440989 5.55547854119109e-10
#> 63  02301500 393913 3081151     5973 47842.6246771642 4.36887888612126e-10
#> 64  02301500 393913 3081151     6880 44804.5783932848 4.98144351268841e-10
#> 65  02301500 393913 3081151     7205 19007.7639926426 2.76782060465964e-09
#> 66  02301500 393913 3081151     8788 46622.2340949037 4.60059325791008e-10
#> 67  02301500 393913 3081151     9401 40782.7492084582 6.01238804577976e-10
#> 68  02301695 374357 3086056      478 43629.3590372355 5.25342255620121e-10
#> 69  02301695 374357 3086056      520   29066.13954415 1.18365542120314e-09
#> 70  02301695 374357 3086056     1632 47266.4924127018 4.47603269241461e-10
#> 71  02301695 374357 3086056     3153 39019.4667826202 6.56806344776614e-10
#> 72  02301695 374357 3086056     3986 28569.0398158566 1.22520486148155e-09
#> 73  02301695 374357 3086056     6880  37191.436971432 7.22959689624267e-10
#> 74  02301695 374357 3086056     7205 19498.5693064902  2.6302347243825e-09
#> 75  02301695 374357 3086056     7851  48633.668379015 4.22791202408455e-10
#> 76  02301695 374357 3086056     7886 37256.5126118911 7.20436319980962e-10
#> 77  02301695 374357 3086056     8788 26488.6414336409 1.42521580885006e-09
#> 78  02301695 374357 3086056     2806 37256.5126118911 7.20436319980962e-10
#> 79  02301750 371419 3090700      478 46780.7607676489 4.56946588817359e-10
#> 80  02301750 371419 3090700      520 23837.0373998112 1.75993020201295e-09
#> 81  02301750 371419 3090700     1632 43726.7699355898  5.2300423429092e-10
#> 82  02301750 371419 3090700     3153 44415.1588537067 5.06917822359547e-10
#> 83  02301750 371419 3090700     3986 24720.2008082459 1.63642464465464e-09
#> 84  02301750 371419 3090700     6880 40611.8700382043 6.06309013435631e-10
#> 85  02301750 371419 3090700     7205 19237.1678009004 2.70220151735371e-09
#> 86  02301750 371419 3090700     7851  44305.017774514 5.09441323210666e-10
#> 87  02301750 371419 3090700     7886 36358.6730230904 7.56456486220407e-10
#> 88  02301750 371419 3090700     8788  22527.090912943 1.97056051559322e-09
#> 89  02301750 371419 3090700     2806 36358.6730230904 7.56456486220407e-10
#> 90  02303000 391854 3115138      478 38966.4015403014 6.58596467478619e-10
#> 91  02303000 391854 3115138      520 40467.0690438534 6.10655818108768e-10
#> 92  02303000 391854 3115138     1641 47578.4484404441 4.41752940765672e-10
#> 93  02303000 391854 3115138     3986 12986.5786487435 5.92939659940211e-09
#> 94  02303000 391854 3115138     4707 37120.0686556477  7.2574233619885e-10
#> 95  02303000 391854 3115138     7205 16080.0396143791 3.86745947213804e-09
#> 96  02303000 391854 3115138     7851 24193.1504562758 1.70850064616092e-09
#> 97  02303000 391854 3115138     8788 47497.4023921309 4.43261775927046e-10
#> 98  02303330 377903 3106951      478 45120.7536838648 4.91187503568262e-10
#> 99  02303330 377903 3106951      520 25360.2443403056 1.55486659382231e-09
#> 100 02303330 377903 3106951     3986 7378.26246483547 1.83692658522677e-08
#> 101 02303330 377903 3106951     7205 13079.3773552108 5.84555646985255e-09
#> 102 02303330 377903 3106951     7851 27642.4072215138  1.3087246388485e-09
#> 103 02303330 377903 3106951     8788 31392.6063269681 1.01471773801655e-09
#> 104 02304500 365288 3121936      520 20906.2381360206  2.2879588753486e-09
#> 105 02304500 365288 3121936     1046 44183.7198637688 5.12242309200501e-10
#> 106 02304500 365288 3121936     1632 46361.0210413878 4.65258178053292e-10
#> 107 02304500 365288 3121936     3986 15623.5449562511 4.09676296855001e-09
#> 108 02304500 365288 3121936     7205 32412.4568954592 9.51866616240835e-10
#> 109 02304500 365288 3121936     7851 17564.1816490265 3.24148607394212e-09
#> 110 02304500 365288 3121936     8788 32001.2137582311 9.76488422291386e-10
#> 111 02304500 365288 3121936     8824 41390.2161023593 5.83720051613064e-10
#> 112 02306647 352788 3104690      520 662.725433343251 2.27684111064309e-06
#> 113 02306647 352788 3104690     1632 26899.4367970781 1.38201776528557e-09
#> 114 02306647 352788 3104690     3986 27808.3272600133 1.29315406243734e-09
#> 115 02306647 352788 3104690     7205 36180.5570161655 7.63922853429454e-10
#> 116 02306647 352788 3104690     7851 38770.6301341621 6.65264395320125e-10
#> 117 02306647 352788 3104690     7886 33524.6802967605 8.89755559047925e-10
#> 118 02306647 352788 3104690     8788 11013.1242615345 8.24477714392114e-09
#> 119 02306647 352788 3104690     8824 29167.5385488731 1.17543992793472e-09
#> 120 02306647 352788 3104690     2806 33524.6802967605 8.89755559047925e-10
#> 121 02307000 350376 3110239      520   5392.697469727 3.43864930405521e-08
#> 122 02307000 350376 3110239     1632  27427.085754779 1.32935407497732e-09
#> 123 02307000 350376 3110239     3986 28803.3326196814 1.20535374257576e-09
#> 124 02307000 350376 3110239     7205 39712.2341980403 6.34090662580959e-10
#> 125 02307000 350376 3110239     7851 36428.0365927125 7.53578452057169e-10
#> 126 02307000 350376 3110239     7886 37865.9523318772 6.97432581748413e-10
#> 127 02307000 350376 3110239     8788 15998.8413642989 3.90681580154626e-09
#> 128 02307000 350376 3110239     8824 25679.3078177742 1.51646843012231e-09
#> 129 02307000 350376 3110239     2806 37865.9523318772 6.97432581748413e-10
#> 130 02307359 340627 3112443      520 13925.2714515732 5.15694690899951e-09
#> 131 02307359 340627 3112443     1632  21960.573990677 2.07354099279608e-09
#> 132 02307359 340627 3112443     3986 38313.2047863397 6.81244545577226e-10
#> 133 02307359 340627 3112443     7205 49695.5189931648 4.04916560366891e-10
#> 134 02307359 340627 3112443     7851 42986.3068080988  5.4117749972957e-10
#> 135 02307359 340627 3112443     7886 38374.5590854149 6.79067902026541e-10
#> 136 02307359 340627 3112443     8788 20071.1152654754 2.48231556194307e-09
#> 137 02307359 340627 3112443     8824 15785.1016151306 4.01333336902525e-09
#> 138 02307359 340627 3112443     2806 38374.5590854149 6.79067902026541e-10
#> 139    202-7 357888 3044944      940 14903.9887614021 4.50189095401782e-09
#> 140    202-7 357888 3044944      945 8466.67969159103 1.39499849793537e-08
#> 141    202-7 357888 3044944     3153 30578.9692599342 1.06943480769313e-09
#> 142    202-7 357888 3044944     6065 32503.0491031226 9.46567942520449e-10
#> 143    202-7 357888 3044944     6880 7109.19102289424 1.97860741534177e-08
#> 144    202-7 357888 3044944     7886  34700.890838709 8.30460062625409e-10
#> 145    202-7 357888 3044944     9176 46800.5335439672 4.56560559283763e-10
#> 146    202-7 357888 3044944     2806  34700.890838709 8.30460062625409e-10
#> 147    203-3 359260 3060684      520 45134.8183224437  4.9088142951741e-10
#> 148    203-3 359260 3060684      940 25328.3747800762   1.558781893568e-09
#> 149    203-3 359260 3060684      945 24088.6815122787  1.7233517944005e-09
#> 150    203-3 359260 3060684     1632 46196.3782130158 4.68580426600868e-10
#> 151    203-3 359260 3060684     3153 30934.4707567464 1.04499597922645e-09
#> 152    203-3 359260 3060684     6065 47198.3498122551 4.48896656347243e-10
#> 153    203-3 359260 3060684     6880 9051.55704837571  1.2205439268572e-08
#> 154    203-3 359260 3060684     7205 48560.7380709972 4.24062082872127e-10
#> 155    203-3 359260 3060684     7886  24240.148039977 1.70188207581831e-09
#> 156    203-3 359260 3060684     8788 35080.6750505175 8.12576234581892e-10
#> 157    203-3 359260 3060684     2806  24240.148039977 1.70188207581831e-09
#> 158    204-2 376572 3079095      478 42024.6408669961  5.6622883060089e-10
#> 159    204-2 376572 3079095      520 35536.6989181606 7.91855312007345e-10
#> 160    204-2 376572 3079095      945 47306.1851769935 4.46852451674476e-10
#> 161    204-2 376572 3079095     3153 31726.8783525893 9.93448396579752e-10
#> 162    204-2 376572 3079095     3986 35243.7726981661 8.05072916444282e-10
#> 163    204-2 376572 3079095     6880 31950.2099210631   9.796085526151e-10
#> 164    204-2 376572 3079095     7205 23585.1234679829 1.79772679568004e-09
#> 165    204-2 376572 3079095     7886  37832.195627534 6.98677739661755e-10
#> 166    204-2 376572 3079095     8788 31322.7161817107 1.01925105434565e-09
#> 167    204-2 376572 3079095     2806  37832.195627534 6.98677739661755e-10
#> 168    205-2 356097 3099401      520 6880.83810302204 2.11211383245948e-08
#> 169    205-2 356097 3099401     1632  28643.001413958 1.21888561359427e-09
#> 170    205-2 356097 3099401     3986 27215.6080402404 1.35009371898316e-09
#> 171    205-2 356097 3099401     6880 47896.2381508193 4.35910360158995e-10
#> 172    205-2 356097 3099401     7205 32482.0005541531 9.47795104441144e-10
#> 173    205-2 356097 3099401     7851 41170.7190245689 5.89960719264888e-10
#> 174    205-2 356097 3099401     7886 30496.3157938791 1.07523959265464e-09
#> 175    205-2 356097 3099401     8788 8604.19548824874 1.35076395156809e-08
#> 176    205-2 356097 3099401     8824 34107.7966013637 8.59592605935348e-10
#> 177    205-2 356097 3099401     2806 30496.3157938791 1.07523959265464e-09
#> 178    206-1 340133 3093481      520 17191.8030758847 3.38342959872867e-09
#> 179    206-1 340133 3093481     1632  12317.412147038 6.59114784325581e-09
#> 180    206-1 340133 3093481     3986   43982.92460717  5.1693006521951e-10
#> 181    206-1 340133 3093481     6880 46355.7112446784 4.65364769787667e-10
#> 182    206-1 340133 3093481     7205 48805.6381169225 4.19816988743487e-10
#> 183    206-1 340133 3093481     7886 19410.0748324163 2.65427294613322e-09
#> 184    206-1 340133 3093481     8788 9081.65909952581 1.21246611096597e-08
#> 185    206-1 340133 3093481     8824 24855.0459464472 1.61871674517424e-09
#> 186    206-1 340133 3093481     2806 19410.0748324163 2.65427294613322e-09
#> 187    206-2 360934 3079741      520 26907.7834464305 1.38116050934493e-09
#> 188    206-2 360934 3079741      940 42637.0544010723 5.50079693125589e-10
#> 189    206-2 360934 3079741      945 43177.9696141447 5.36383694480654e-10
#> 190    206-2 360934 3079741     1632 36243.3919632255 7.61276332787201e-10
#> 191    206-2 360934 3079741     3153 40631.0599664838 6.05736432808964e-10
#> 192    206-2 360934 3079741     3986 38914.8918795877 6.60341122582946e-10
#> 193    206-2 360934 3079741     6880 28073.4796204532 1.26884189290818e-09
#> 194    206-2 360934 3079741     7205 33919.4006580305 8.69167863320322e-10
#> 195    206-2 360934 3079741     7886 22579.0090570866  1.9615087258206e-09
#> 196    206-2 360934 3079741     8788 18700.4802344753 2.85952883983673e-09
#> 197    206-2 360934 3079741     8824 49128.2328809006 4.14321732374822e-10
#> 198    206-2 360934 3079741     2806 22579.0090570866  1.9615087258206e-09
#> 199   206-3C 348225 3071277      520 34326.3034566788 8.48683824927937e-10
#> 200   206-3C 348225 3071277      940  30831.135123443 1.05201267103395e-09
#> 201   206-3C 348225 3071277      945 34985.3670125097  8.1700954725902e-10
#> 202   206-3C 348225 3071277     1632 30939.0909045499 1.04468390312786e-09
#> 203   206-3C 348225 3071277     3153  45429.665693245 4.84530274729514e-10
#> 204   206-3C 348225 3071277     6880 22950.6462871964 1.89849807254509e-09
#> 205   206-3C 348225 3071277     7205 49184.0140696141 4.13382473816643e-10
#> 206   206-3C 348225 3071277     7886 9579.40734074922 1.08973956259707e-08
#> 207   206-3C 348225 3071277     8788 23027.7697574038 1.88580265550432e-09
#> 208   206-3C 348225 3071277     8824 47897.2794738908 4.35891406323379e-10
#> 209   206-3C 348225 3071277     2806 9579.40734074922 1.08973956259707e-08
#> 210   206-3E 355763 3065892      520 39559.0551959978 6.39010771247353e-10
#> 211   206-3E 355763 3065892      940 27897.6545609125 1.28488605367074e-09
#> 212   206-3E 355763 3065892      945 28935.6385448809 1.19435617309641e-09
#> 213   206-3E 355763 3065892     1632 40011.2059428356 6.24649961387481e-10
#> 214   206-3E 355763 3065892     3153 36251.4529363997 7.60937811413799e-10
#> 215   206-3E 355763 3065892     6880 14880.6253900836 4.51603846674858e-09
#> 216   206-3E 355763 3065892     7205 46897.1306691572 4.54681680276968e-10
#> 217   206-3E 355763 3065892     7886 18600.9173967307 2.89022246065401e-09
#> 218   206-3E 355763 3065892     8788  29146.380650091  1.1771470946706e-09
#> 219   206-3E 355763 3065892     2806 18600.9173967307 2.89022246065401e-09
#> 220   206-3W 337811 3075646      520 33156.4524640378 9.09628135467717e-10
#> 221   206-3W 337811 3075646      940 35320.0018403171  8.0160158071984e-10
#> 222   206-3W 337811 3075646      945 42351.0305187489 5.57534870630997e-10
#> 223   206-3W 337811 3075646     1632 21380.7456605236  2.1875314491826e-09
#> 224   206-3W 337811 3075646     6880 32775.8411028611 9.30877019453915e-10
#> 225   206-3W 337811 3075646     7886 1994.42121930148 2.51400552477854e-07
#> 226   206-3W 337811 3075646     8788 21832.1722464807 2.09800298543727e-09
#> 227   206-3W 337811 3075646     8824 39618.3232355939 6.37100312443168e-10
#> 228   206-3W 337811 3075646     2806 1994.42121930148 2.51400552477854e-07
#> 229    206-4 339637 3053478      940 13275.5891018064 5.67404004008913e-09
#> 230    206-4 339637 3053478      945 22592.5813044902 1.95915272365428e-09
#> 231    206-4 339637 3053478     1632 42747.1679178867 5.47249414581991e-10
#> 232    206-4 339637 3053478     3153 48641.3474319945 4.22657720096007e-10
#> 233    206-4 339637 3053478     6065 49079.7040129624 4.15141480802007e-10
#> 234    206-4 339637 3053478     6880 20608.7578470902 2.35448738218439e-09
#> 235    206-4 339637 3053478     7886 20629.7658736109 2.34969450681839e-09
#> 236    206-4 339637 3053478     8788 41907.7679314945 5.69391447283696e-10
#> 237    206-4 339637 3053478     2806 20629.7658736109 2.34969450681839e-09
#> 238    206-5 331790 3062133      520 47940.6978046837 4.35102218516335e-10
#> 239    206-5 331790 3062133      940 24371.1936515223 1.68362900440215e-09
#> 240    206-5 331790 3062133      945  34274.407069999 8.51255829634016e-10
#> 241    206-5 331790 3062133     1632 32675.7126471635 9.36590748628009e-10
#> 242    206-5 331790 3062133     6880 30241.7121539109 1.09342059372008e-09
#> 243    206-5 331790 3062133     7886 14007.9281123227 5.09626721359806e-09
#> 244    206-5 331790 3062133     8788 36553.0904986158 7.48431049956893e-10
#> 245    206-5 331790 3062133     2806 14007.9281123227 5.09626721359806e-09
#> 246    206-6 344186 3048396      940 7621.77177564377 1.72142505418831e-08
#> 247    206-6 344186 3048396      945 15774.1770308311 4.01889425260786e-09
#> 248    206-6 344186 3048396     1632 48975.8170937454 4.16904535023329e-10
#> 249    206-6 344186 3048396     3153 43947.9964617274 5.17752063254528e-10
#> 250    206-6 344186 3048396     6065 42279.7641076674 5.59416006789381e-10
#> 251    206-6 344186 3048396     6880 16314.5298737046 3.75708372306032e-09
#> 252    206-6 344186 3048396     7886 26207.9919299438 1.45590322247887e-09
#> 253    206-6 344186 3048396     8788 46159.8350300345 4.69322638884619e-10
#> 254    206-6 344186 3048396     2806 26207.9919299438 1.45590322247887e-09
#> 255    207-5 327889 3078015      520 36826.1852083541 7.37371816177539e-10
#> 256    207-5 327889 3078015      940 40385.4818592028 6.13125617246971e-10
#> 257    207-5 327889 3078015      945 49225.4851575888 4.12686240982826e-10
#> 258    207-5 327889 3078015     1632 16557.0188741814 3.64783942082466e-09
#> 259    207-5 327889 3078015     6880 41659.4149382826 5.76200548140388e-10
#> 260    207-5 327889 3078015     7886 11847.5461172346 7.12431616093628e-09
#> 261    207-5 327889 3078015     8788 26795.0928343232 1.39280227639604e-09
#> 262    207-5 327889 3078015     8824 35201.9913783297 8.06985136157724e-10
#> 263    207-5 327889 3078015     2806 11847.5461172346 7.12431616093628e-09
#> 264 EVERSRES 360470 3032755      940 18714.9229226305 2.85511702924845e-09
#> 265 EVERSRES 360470 3032755      945 6858.24292366492 2.12605390618183e-08
#> 266 EVERSRES 360470 3032755     3153 32347.1439542968  9.5571437306931e-10
#> 267 EVERSRES 360470 3032755     6065  20304.684878126 2.42553470209103e-09
#> 268 EVERSRES 360470 3032755     6880 18925.4257019492 2.79195674562129e-09
#> 269 EVERSRES 360470 3032755     7886  46554.551077204 4.61398006448523e-10
#> 270 EVERSRES 360470 3032755     9176 34839.0376589251 8.23887093622892e-10
#> 271 EVERSRES 360470 3032755     2806  46554.551077204 4.61398006448523e-10
#> 272 LMANATEE 374561 3041023      940 30989.8095670173 1.04126719210099e-09
#> 273 LMANATEE 374561 3041023      945 19922.6601888402 2.51944768597284e-09
#> 274 LMANATEE 374561 3041023     3153 16013.3321017207  3.8997483121938e-09
#> 275 LMANATEE 374561 3041023     6065 26833.5977647426 1.38880793488858e-09
#> 276 LMANATEE 374561 3041023     6880 17908.5527332613 3.11802088513373e-09
#> 277 LMANATEE 374561 3041023     7886 48514.2044044834 4.24875973466749e-10
#> 278 LMANATEE 374561 3041023     9176 46597.7944757045 4.60542036331286e-10
#> 279 LMANATEE 374561 3041023     9401 45577.4947205307 4.81392257655193e-10
#> 280 LMANATEE 374561 3041023     2806 48514.2044044834 4.24875973466749e-10
#> 281  LTARPON 331573 3109280      520 21390.9892244375 2.18543685199885e-09
#> 282  LTARPON 331573 3109280     1632 15168.4522941532 4.34627758384084e-09
#> 283  LTARPON 331573 3109280     3986 47585.5156534002 4.41621735703258e-10
#> 284  LTARPON 331573 3109280     7886 35970.1599106815 7.72885683798598e-10
#> 285  LTARPON 331573 3109280     8788 23126.2029092543 1.86978355746407e-09
#> 286  LTARPON 331573 3109280     8824 7721.21493030728 1.67736940169575e-08
#> 287  LTARPON 331573 3109280     2806 35970.1599106815 7.72885683798598e-10
#> 288  TBYPASS 368888 3097382      520 18125.5955212512 3.04379522450018e-09
#> 289  TBYPASS 368888 3097382     1632  41120.124951172 5.91413384047772e-10
#> 290  TBYPASS 368888 3097382     3986 19624.3338995238 2.59663048089848e-09
#> 291  TBYPASS 368888 3097382     6880  46528.609016389 4.61912655488339e-10
#> 292  TBYPASS 368888 3097382     7205 19793.6264994568 2.55240306834583e-09
#> 293  TBYPASS 368888 3097382     7851 38184.6674072199 6.85838681174738e-10
#> 294  TBYPASS 368888 3097382     7886 37829.9486253947 6.98760741466287e-10
#> 295  TBYPASS 368888 3097382     8788  19950.841836875 2.51233499957253e-09
#> 296  TBYPASS 368888 3097382     8824 46749.5529925153 4.57556864006008e-10
#> 297  TBYPASS 368888 3097382     2806 37829.9486253947 6.98760741466287e-10