Add column names for DPS reuse and end of pipe from raw entity data

Add column names for DPS reuse and end of pipe from raw entity data

Usage

util_dps_addcol(dat)

Arguments

dat

data frame from raw entity data as data.frame

Value

Input data frame from pth as is if column names are correct, otherwise additional columns are added as needed.

Details

The function checks for TN, TP, TSS, and BOD. If any of these are missing, the columns are added with empty values including a column for units. If BOD is missing but CBOD is present, the CBOD column is renamed to BOD.

Examples

pth <- system.file('extdata/ps_dom_hillsco_falkenburg_2019.txt', package = 'tbeploads')
dat <- read.table(pth, skip = 0, sep = '\t', header = TRUE)
util_dps_addcol(dat)
#>    Permit.Number         Facility.Name Outfall.ID Year Month
#> 1        FL00407 Falkenburg Road AWWTF      D-001 2019     1
#> 2        FL00407 Falkenburg Road AWWTF      D-001 2019     2
#> 3        FL00407 Falkenburg Road AWWTF      D-001 2019     3
#> 4        FL00407 Falkenburg Road AWWTF      D-001 2019     4
#> 5        FL00407 Falkenburg Road AWWTF      D-001 2019     5
#> 6        FL00407 Falkenburg Road AWWTF      D-001 2019     6
#> 7        FL00407 Falkenburg Road AWWTF      D-001 2019     7
#> 8        FL00407 Falkenburg Road AWWTF      D-001 2019     8
#> 9        FL00407 Falkenburg Road AWWTF      D-001 2019     9
#> 10       FL00407 Falkenburg Road AWWTF      D-001 2019    10
#> 11       FL00407 Falkenburg Road AWWTF      D-001 2019    11
#> 12       FL00407 Falkenburg Road AWWTF      D-001 2019    12
#> 13       FL00407 Falkenburg Road AWWTF      R-001 2019     1
#> 14       FL00407 Falkenburg Road AWWTF      R-001 2019     2
#> 15       FL00407 Falkenburg Road AWWTF      R-001 2019     3
#> 16       FL00407 Falkenburg Road AWWTF      R-001 2019     4
#> 17       FL00407 Falkenburg Road AWWTF      R-001 2019     5
#> 18       FL00407 Falkenburg Road AWWTF      R-001 2019     6
#> 19       FL00407 Falkenburg Road AWWTF      R-001 2019     7
#> 20       FL00407 Falkenburg Road AWWTF      R-001 2019     8
#> 21       FL00407 Falkenburg Road AWWTF      R-001 2019     9
#> 22       FL00407 Falkenburg Road AWWTF      R-001 2019    10
#> 23       FL00407 Falkenburg Road AWWTF      R-001 2019    11
#> 24       FL00407 Falkenburg Road AWWTF      R-001 2019    12
#>    Average.Daily.Flow..ADF...mgd.  BOD BOD..Unit QBOD  TSS TSS.Unit QTSS
#> 1                           3.703 1.48      mg/l   NA 0.64     mg/l   NA
#> 2                           3.691 1.02      mg/l   NA 0.51     mg/l   NA
#> 3                           3.542 0.99      mg/l   NA 0.44     mg/l   NA
#> 4                           2.825 1.77      mg/l   NA 0.36     mg/l   NA
#> 5                           4.433 1.30      mg/l   NA 0.37     mg/l   NA
#> 6                           4.962 1.06      mg/l   NA 0.35     mg/l   NA
#> 7                           6.070 1.10      mg/l   NA 0.35     mg/l   NA
#> 8                           7.400 1.19      mg/l   NA 0.30     mg/l   NA
#> 9                           4.974 1.41      mg/l   NA 0.28     mg/l   NA
#> 10                          5.467 1.40      mg/l   NA 0.29     mg/l   NA
#> 11                          7.176 1.56      mg/l   NA 0.31     mg/l   NA
#> 12                          7.389 1.17      mg/l   NA 0.50     mg/l   NA
#> 13                          4.240 1.48      mg/l   NA 0.64     mg/l   NA
#> 14                          4.347 1.02      mg/l   NA 0.51     mg/l   NA
#> 15                          4.410 0.99      mg/l   NA 0.44     mg/l   NA
#> 16                          5.077 1.77      mg/l   NA 0.36     mg/l   NA
#> 17                          5.357 1.30      mg/l   NA 0.37     mg/l   NA
#> 18                          4.777 1.06      mg/l   NA 0.35     mg/l   NA
#> 19                          4.058 1.10      mg/l   NA 0.35     mg/l   NA
#> 20                          3.297 1.22      mg/l   NA 0.29     mg/l   NA
#> 21                          4.756 1.44      mg/l   NA 0.28     mg/l   NA
#> 22                          4.211 1.40      mg/l   NA 0.29     mg/l   NA
#> 23                          2.508 1.58      mg/l   NA 0.32     mg/l   NA
#> 24                          2.733 1.09      mg/l   NA 0.47     mg/l   NA
#>    Total.N Total.N.Unit Q.Total.N NO2.NO3 NO2.NO3.Unit Q.NO2.NO3  NH4 NH4.Unit
#> 1     2.36         mg/l        NA    1.24         mg/l        NA 0.27     mg/l
#> 2     2.42         mg/l        NA    1.28         mg/l        NA 0.35     mg/l
#> 3     2.41         mg/l        NA    1.25         mg/l        NA 0.39     mg/l
#> 4     1.99         mg/l        NA    1.07         mg/l        NA 0.29     mg/l
#> 5     1.84         mg/l        NA    0.94         mg/l        NA 0.17     mg/l
#> 6     1.84         mg/l        NA    1.02         mg/l        NA 0.13     mg/l
#> 7     1.97         mg/l        NA    1.19         mg/l        NA 0.05     mg/l
#> 8     1.89         mg/l        NA    1.13         mg/l        NA 0.07     mg/l
#> 9     2.11         mg/l        NA    1.33         mg/l        NA 0.07     mg/l
#> 10    1.80         mg/l        NA    1.10         mg/l        NA 0.05     mg/l
#> 11    1.67         mg/l        NA    0.83         mg/l        NA 0.05     mg/l
#> 12    1.91         mg/l        NA    0.93         mg/l        NA 0.20     mg/l
#> 13    2.36         mg/l        NA    1.24         mg/l        NA 0.27     mg/l
#> 14    2.42         mg/l        NA    1.28         mg/l        NA 0.35     mg/l
#> 15    2.41         mg/l        NA    1.25         mg/l        NA 0.39     mg/l
#> 16    1.99         mg/l        NA    1.07         mg/l        NA 0.29     mg/l
#> 17    1.84         mg/l        NA    0.94         mg/l        NA 0.17     mg/l
#> 18    1.84         mg/l        NA    1.02         mg/l        NA 0.13     mg/l
#> 19    1.97         mg/l        NA    1.19         mg/l        NA 0.05     mg/l
#> 20    1.86         mg/l        NA    1.10         mg/l        NA 0.07     mg/l
#> 21    2.09         mg/l        NA    1.32         mg/l        NA 0.07     mg/l
#> 22    1.80         mg/l        NA    1.10         mg/l        NA 0.05     mg/l
#> 23    1.65         mg/l        NA    0.81         mg/l        NA 0.05     mg/l
#> 24    1.93         mg/l        NA    0.98         mg/l        NA 0.19     mg/l
#>    Q.NH4  TKN TKN.Unit Q.TKN Organic.N Organic.N.Unit Q.Organic.N Total.P
#> 1     NA 1.12     mg/l    NA      0.84           MG/L          NA    0.28
#> 2     NA 1.14     mg/l    NA      0.79           MG/L          NA    0.41
#> 3     NA 1.16     mg/l    NA      0.77           MG/L          NA    0.22
#> 4     NA 0.92     mg/l    NA      0.63           MG/L          NA    0.19
#> 5     NA 0.89     mg/l    NA      0.73           MG/L          NA    0.21
#> 6     NA 0.83     mg/l    NA      0.70           MG/L          NA    0.18
#> 7     NA 0.78     mg/l    NA      0.73           MG/L          NA    0.20
#> 8     NA 0.77     mg/l    NA      0.69           MG/L          NA    0.24
#> 9     NA 0.77     mg/l    NA      0.70           MG/L          NA    0.20
#> 10    NA 0.70     mg/l    NA      0.65           MG/L          NA    0.30
#> 11    NA 0.81     mg/l    NA      0.76           MG/L          NA    0.35
#> 12    NA 0.98     mg/l    NA      0.77           MG/L          NA    0.26
#> 13    NA 1.12     mg/l    NA      0.84           MG/L          NA    0.28
#> 14    NA 1.14     mg/l    NA      0.79           MG/L          NA    0.41
#> 15    NA 1.16     mg/l    NA      0.77           MG/L          NA    0.22
#> 16    NA 0.92     mg/l    NA      0.63           MG/L          NA    0.19
#> 17    NA 0.89     mg/l    NA      0.73           MG/L          NA    0.21
#> 18    NA 0.83     mg/l    NA      0.70           MG/L          NA    0.18
#> 19    NA 0.78     mg/l    NA      0.73           MG/L          NA    0.20
#> 20    NA 0.76     mg/l    NA      0.69           MG/L          NA    0.24
#> 21    NA 0.77     mg/l    NA      0.70           MG/L          NA    0.20
#> 22    NA 0.70     mg/l    NA      0.65           MG/L          NA    0.30
#> 23    NA 0.81     mg/l    NA      0.76           MG/L          NA    0.36
#> 24    NA 0.95     mg/l    NA      0.76           MG/L          NA    0.26
#>    Total.P.Unit Q.Total.P  PO4 PO4.Unit Q.PO4 Other.N..species. Other.N.Unit
#> 1          mg/l        NA 0.28     mg/l    NA                NA           NA
#> 2          mg/l        NA 0.45     mg/l    NA                NA           NA
#> 3          mg/l        NA 0.23     mg/l    NA                NA           NA
#> 4          mg/l        NA 0.20     mg/l    NA                NA           NA
#> 5          mg/l        NA 0.22     mg/l    NA                NA           NA
#> 6          mg/l        NA 0.19     mg/l    NA                NA           NA
#> 7          mg/l        NA   NA     mg/l    NA                NA           NA
#> 8          mg/l        NA   NA     mg/l    NA                NA           NA
#> 9          mg/l        NA   NA     mg/l    NA                NA           NA
#> 10         mg/l        NA 0.31     mg/l    NA                NA           NA
#> 11         mg/l        NA 0.35     mg/l    NA                NA           NA
#> 12         mg/l        NA 0.28     mg/l    NA                NA           NA
#> 13         mg/l        NA 0.28     mg/l    NA                NA           NA
#> 14         mg/l        NA 0.45     mg/l    NA                NA           NA
#> 15         mg/l        NA 0.23     mg/l    NA                NA           NA
#> 16         mg/l        NA 0.20     mg/l    NA                NA           NA
#> 17         mg/l        NA 0.22     mg/l    NA                NA           NA
#> 18         mg/l        NA 0.19     mg/l    NA                NA           NA
#> 19         mg/l        NA   NA     mg/l    NA                NA           NA
#> 20         mg/l        NA   NA     mg/l    NA                NA           NA
#> 21         mg/l        NA   NA     mg/l    NA                NA           NA
#> 22         mg/l        NA 0.31     mg/l    NA                NA           NA
#> 23         mg/l        NA 0.35     mg/l    NA                NA           NA
#> 24         mg/l        NA 0.28     mg/l    NA                NA           NA
#>    Q.Other.N Other.P..species. Other.P.Unit Q.Other.P
#> 1         NA                NA           NA        NA
#> 2         NA                NA           NA        NA
#> 3         NA                NA           NA        NA
#> 4         NA                NA           NA        NA
#> 5         NA                NA           NA        NA
#> 6         NA                NA           NA        NA
#> 7         NA                NA           NA        NA
#> 8         NA                NA           NA        NA
#> 9         NA                NA           NA        NA
#> 10        NA                NA           NA        NA
#> 11        NA                NA           NA        NA
#> 12        NA                NA           NA        NA
#> 13        NA                NA           NA        NA
#> 14        NA                NA           NA        NA
#> 15        NA                NA           NA        NA
#> 16        NA                NA           NA        NA
#> 17        NA                NA           NA        NA
#> 18        NA                NA           NA        NA
#> 19        NA                NA           NA        NA
#> 20        NA                NA           NA        NA
#> 21        NA                NA           NA        NA
#> 22        NA                NA           NA        NA
#> 23        NA                NA           NA        NA
#> 24        NA                NA           NA        NA