Check units for DPS reuse and end of pipe from raw entity data

Check units for DPS reuse and end of pipe from raw entity data

Usage

util_dps_checkuni(dat)

Arguments

dat

data frame from raw entity data as data.frame

Value

Input data frame from pth with relevant data and columns renamed, otherwise an error is returned if units are not correct. Only year, month, outfall, flow, TN, TP, TSS, and BOD are returned.

Details

Input data should include flow as million gallons per day, and concentration as mg/L.

Examples

pth <- system.file('extdata/ps_dom_hillsco_falkenburg_2019.txt', package = 'tbeploads')
dat <- read.table(pth, skip = 0, sep = '\t', header = TRUE)
util_dps_checkuni(dat)
#>    Year Month outfall flow_mgd tn_mgl tp_mgl tss_mgl bod_mgl
#> 1  2019     1   D-001    3.703   2.36   0.28    0.64    1.48
#> 2  2019     2   D-001    3.691   2.42   0.41    0.51    1.02
#> 3  2019     3   D-001    3.542   2.41   0.22    0.44    0.99
#> 4  2019     4   D-001    2.825   1.99   0.19    0.36    1.77
#> 5  2019     5   D-001    4.433   1.84   0.21    0.37    1.30
#> 6  2019     6   D-001    4.962   1.84   0.18    0.35    1.06
#> 7  2019     7   D-001    6.070   1.97   0.20    0.35    1.10
#> 8  2019     8   D-001    7.400   1.89   0.24    0.30    1.19
#> 9  2019     9   D-001    4.974   2.11   0.20    0.28    1.41
#> 10 2019    10   D-001    5.467   1.80   0.30    0.29    1.40
#> 11 2019    11   D-001    7.176   1.67   0.35    0.31    1.56
#> 12 2019    12   D-001    7.389   1.91   0.26    0.50    1.17
#> 13 2019     1   R-001    4.240   2.36   0.28    0.64    1.48
#> 14 2019     2   R-001    4.347   2.42   0.41    0.51    1.02
#> 15 2019     3   R-001    4.410   2.41   0.22    0.44    0.99
#> 16 2019     4   R-001    5.077   1.99   0.19    0.36    1.77
#> 17 2019     5   R-001    5.357   1.84   0.21    0.37    1.30
#> 18 2019     6   R-001    4.777   1.84   0.18    0.35    1.06
#> 19 2019     7   R-001    4.058   1.97   0.20    0.35    1.10
#> 20 2019     8   R-001    3.297   1.86   0.24    0.29    1.22
#> 21 2019     9   R-001    4.756   2.09   0.20    0.28    1.44
#> 22 2019    10   R-001    4.211   1.80   0.30    0.29    1.40
#> 23 2019    11   R-001    2.508   1.65   0.36    0.32    1.58
#> 24 2019    12   R-001    2.733   1.93   0.26    0.47    1.09