Utility function for non-point source (NPS) ungaged workflow to prepare land use and soil data for logistic regression

Utility function for non-point source (NPS) ungaged workflow to prepare land use and soil data for logistic regression

Usage

util_nps_preplog(tbbase)

Arguments

tbbase

Input data frame returned from util_nps_tbbase.

Value

A data frame with land use and soil areas in long format with bay segment and basins in the rows.

Examples

util_nps_preplog(tbbase)
#> # A tibble: 71 × 190
#> # Groups:   bay_seg [8]
#>    bay_seg basin    `C_C01 `   C_C01A   C_C01B   C_C01C C_C01D `C_C02 `   C_C02A
#>      <dbl> <chr>       <dbl>    <dbl>    <dbl>    <dbl>  <dbl>    <dbl>    <dbl>
#>  1       1 02304500   NA      1.95e-5  7.57e-3  4.04e-3   NA      NA     1.08e-3
#>  2       1 02306647    0.869  3.18e+0  9.80e+1  8.68e+1   22.0     2.70  1.84e+1
#>  3       1 02307000   23.0    1.79e+1  7.73e+2  2.48e+2  106.      6.43  4.96e+1
#>  4       1 02307359   43.2    1.23e+1  9.32e+2  5.29e+2  163.      2.50  9.96e+0
#>  5       1 205-2      NA     NA       NA       NA         NA      NA    NA      
#>  6       1 206-1       2.65   5.15e+1  7.51e+2  8.45e+1   37.1     6.89  1.46e+2
#>  7       1 206-2      NA     NA       NA       NA         NA      NA    NA      
#>  8       1 206-3C     NA     NA       NA       NA         NA      NA    NA      
#>  9       1 206-3W     NA     NA       NA       NA         NA      NA    NA      
#> 10       1 207-5      NA     NA       NA       NA         NA      NA    NA      
#> # ℹ 61 more rows
#> # ℹ 181 more variables: C_C02B <dbl>, C_C02C <dbl>, C_C02D <dbl>,
#> #   `C_C03 ` <dbl>, C_C03A <dbl>, C_C03B <dbl>, C_C03C <dbl>, C_C03D <dbl>,
#> #   `C_C04 ` <dbl>, C_C04A <dbl>, C_C04B <dbl>, C_C04C <dbl>, C_C04D <dbl>,
#> #   `C_C05 ` <dbl>, C_C05A <dbl>, C_C05B <dbl>, C_C05C <dbl>, C_C05D <dbl>,
#> #   `C_C06 ` <dbl>, C_C06A <dbl>, C_C06B <dbl>, C_C06C <dbl>, C_C06D <dbl>,
#> #   `C_C07 ` <dbl>, C_C07A <dbl>, C_C07B <dbl>, C_C07C <dbl>, C_C07D <dbl>, …