Light edits to the outfall ID column for point source data

Light edits to the outfall ID column for point source data

Usage

util_ps_fixoutfall(dat)

Arguments

dat

data frame from raw entity data as data.frame

Value

Input data frame as is, with any edits to the outfall ID column.

Details

The outfall ID column is edited lightly to remove any leading or trailing white space, a hyphen is added between letters and numbers if missing, and "Outfall" prefix is removed if presenn.

Examples

pth <- system.file('extdata/ps_ind_busch_busch_2020.txt', package = 'tbeploads')
dat <- read.table(pth, skip = 0, sep = '\t', header = TRUE)
util_ps_fixoutfall(dat)
#>    Permit.Number Facility.Name Outfall.ID Year Month
#> 1      FL0185833            NA      D-002 2020     1
#> 2      FL0185833            NA      D-002 2020     2
#> 3      FL0185833            NA      D-002 2020     3
#> 4      FL0185833            NA      D-002 2020     4
#> 5      FL0185833            NA      D-002 2020     5
#> 6      FL0185833            NA      D-002 2020     6
#> 7      FL0185833            NA      D-002 2020     7
#> 8      FL0185833            NA      D-002 2020     8
#> 9      FL0185833            NA      D-002 2020     9
#> 10     FL0185833            NA      D-002 2020    10
#> 11     FL0185833            NA      D-002 2020    11
#> 12     FL0185833            NA      D-002 2020    12
#>    Average.Daily.Flow..ADF...mgd. BOD BOD..Unit QBOD TSS TSS.Unit QTSS Total.N
#> 1                            0.92 9.6        NA   NA   5       NA   NA    0.72
#> 2                            1.04 9.6        NA   NA   5       NA   NA    0.30
#> 3                            0.28 9.6        NA   NA   5       NA   NA    0.40
#> 4                            0.70 9.6        NA   NA   5       NA   NA    0.56
#> 5                            0.42 9.6        NA   NA   5       NA   NA    0.41
#> 6                            0.57 9.6        NA   NA   5       NA   NA    0.60
#> 7                            0.82 9.6        NA   NA   5       NA   NA    0.54
#> 8                            1.02 9.6        NA   NA   5       NA   NA    0.86
#> 9                            1.03 9.6        NA   NA   5       NA   NA    0.44
#> 10                           0.59 9.6        NA   NA   5       NA   NA    0.59
#> 11                           0.97 9.6        NA   NA   5       NA   NA    0.19
#> 12                           0.84 9.6        NA   NA   5       NA   NA    0.06
#>    Total.N.Unit Q.Total.N NO2.NO3 NO2.NO3.Unit Q.NO2.NO3    NH4 NH4.Unit
#> 1          mg/L   Average    0.06         mg/L   Average < 0.02     mg/L
#> 2          mg/L   Average    0.03         mg/L   Average < 0.02     mg/L
#> 3          mg/L   Average  < 0.01         mg/L         < < 0.02     mg/L
#> 4          mg/L   Average  < 0.01         mg/L         < < 0.02     mg/L
#> 5          mg/L   Average    0.01         mg/L   Average < 0.02     mg/L
#> 6          mg/L   Average    0.03         mg/L   Average < 0.02     mg/L
#> 7          mg/L   Average    0.02         mg/L   Average   0.02     mg/L
#> 8          mg/L   Average    0.01         mg/L   Average < 0.02     mg/L
#> 9          mg/L   Average    0.03         mg/L   Average   0.03     mg/L
#> 10         mg/L   Average    0.05         mg/L   Average < 0.02     mg/L
#> 11         mg/L   Average    0.04         mg/L   Average   0.02     mg/L
#> 12         mg/L   Average    0.04         mg/L   Average < 0.02     mg/L
#>      Q.NH4  TKN TKN.Unit   Q.TKN Organic.N Organic.N.Unit Q.Organic.N Total.P
#> 1        < 0.67     mg/L Average      0.67           mg/L     Average    0.04
#> 2        < 0.39     mg/L Average      0.39           mg/L     Average    0.02
#> 3        < 0.40     mg/L Average      0.40           mg/L     Average    0.04
#> 4        < 0.56     mg/L Average      0.56           mg/L     Average    0.02
#> 5        < 0.41     mg/L Average      0.41           mg/L     Average    0.04
#> 6        < 0.57     mg/L Average      0.57           mg/L     Average    0.07
#> 7  Average 0.52     mg/L Average      0.50           mg/L     Average    0.10
#> 8        < 0.85     mg/L Average      0.85           mg/L     Average    0.12
#> 9  Average 0.41     mg/L Average      0.38           mg/L     Average    0.08
#> 10       < 0.59     mg/L Average      0.59           mg/L     Average    0.13
#> 11 Average 0.17     mg/L Average      0.15           mg/L     Average    0.08
#> 12       < 0.04     mg/L Average      0.04           mg/L     Average    0.08
#>    Total.P.Unit Q.Total.P  PO4 PO4.Unit   Q.PO4 Other.N..species. Other.N.Unit
#> 1          mg/L   Average 0.02     mg/L Average                NA           NA
#> 2          mg/L   Average 0.01     mg/L Average                NA           NA
#> 3          mg/L   Average 0.02     mg/L Average                NA           NA
#> 4          mg/L   Average 0.01     mg/L Average                NA           NA
#> 5          mg/L   Average 0.03     mg/L Average                NA           NA
#> 6          mg/L   Average 0.05     mg/L Average                NA           NA
#> 7          mg/L   Average 0.02     mg/L Average                NA           NA
#> 8          mg/L   Average 0.06     mg/L Average                NA           NA
#> 9          mg/L   Average 0.06     mg/L Average                NA           NA
#> 10         mg/L   Average 0.02     mg/L Average                NA           NA
#> 11         mg/L   Average 0.05     mg/L Average                NA           NA
#> 12         mg/L   Average 0.01     mg/L Average                NA           NA
#>    Q.Other.N Other.P..species. Other.P.Unit Q.Other.P
#> 1         NA                NA           NA        NA
#> 2         NA                NA           NA        NA
#> 3         NA                NA           NA        NA
#> 4         NA                NA           NA        NA
#> 5         NA                NA           NA        NA
#> 6         NA                NA           NA        NA
#> 7         NA                NA           NA        NA
#> 8         NA                NA           NA        NA
#> 9         NA                NA           NA        NA
#> 10        NA                NA           NA        NA
#> 11        NA                NA           NA        NA
#> 12        NA                NA           NA        NA