Calculate material losses (ML) from raw facility data

Calculate material losses (ML) from raw facility data

Usage

anlz_ml_facility(fls)

Arguments

fls

vector of file paths to raw facility data, one to many

Value

data frame that is nearly identical to the input data except results are shown as monthly load as the annual estimate divided by 12. This is for consistency of reporting without sources.

Details

Input data should one row per year per facility, where the row shows the total tons per year of total nitrogen loss.

See also

Examples

fls <- list.files(system.file('extdata/', package = 'tbeploads'),
  pattern = 'ps_indml', full.names = TRUE)
anlz_ml_facility(fls)
#> Called from: anlz_ml_facility(fls)
#> debug: mlprep <- tidyr::unnest(tidyr::unnest(dplyr::ungroup(dplyr::mutate(tidyr::nest(dplyr::group_by(tibble::tibble(fls = fls), 
#>     fls), .key = "dat"), dat = purrr::map(fls, read.table, skip = 0, 
#>     sep = "\t", header = T), entinfo = purrr::map(fls, util_ps_facinfo, 
#>     asdf = T))), "entinfo"), "dat")
#> debug: ml <- dplyr::select(dplyr::mutate(dplyr::full_join(tidyr::crossing(unique(mlprep[, 
#>     c("year", "entity", "facname")]), Month = 1:12), mlprep, 
#>     by = c("year", "entity", "facname")), tn_load = tn_tonsyr/12, 
#>     tp_load = NA, tss_load = NA, bod_load = NA, hy_load = NA, 
#>     source = NA), Year = year, Month, entity, facility = facname, 
#>     coastco, source, tn_load, tp_load, tss_load, bod_load, hy_load)
#> debug: out <- ml
#> debug: return(out)
#> # A tibble: 60 × 11
#>     Year Month entity  facility coastco source tn_load tp_load tss_load bod_load
#>    <int> <int> <chr>   <chr>    <chr>   <lgl>    <dbl> <lgl>   <lgl>    <lgl>   
#>  1  2017     1 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  2  2017     2 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  3  2017     3 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  4  2017     4 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  5  2017     5 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  6  2017     6 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  7  2017     7 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  8  2017     8 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#>  9  2017     9 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#> 10  2017    10 Kinder… Kinder … NA      NA      0.0156 NA      NA       NA      
#> # ℹ 50 more rows
#> # ℹ 1 more variable: hy_load <lgl>