Utility function for non-point source (NPS) ungaged workflow to prepare land use and soil data for logistic regression

Utility function for non-point source (NPS) ungaged workflow to prepare land use and soil data for logistic regression

Usage

util_nps_preplog(tbbase)

Arguments

tbbase

Input data frame returned from util_nps_tbbase.

Value

A data frame with land use and soil areas in long format with bay segment and basins in the rows.

Examples

util_nps_preplog(tbbase)
#> # A tibble: 6 × 190
#>   bay_seg basin   `C_C01 ` C_C01A  C_C01B  C_C01C C_C01D `C_C02 `  C_C02A C_C02B
#>     <dbl> <chr>      <dbl>  <dbl>   <dbl>   <dbl>  <dbl>    <dbl>   <dbl>  <dbl>
#> 1       2 023015…    NA    1017.  1.46e+3 1.20e+3   107.    NA      NA       NA 
#> 2       3 023015…    NA      NA   1.46e-2 3.34e-4    NA     NA      NA       NA 
#> 3       2 023045…     8.41 3968.  7.60e+3 2.79e+3   934.    17.9  3222.    2095.
#> 4       2 023030…     3.53 1492.  3.82e+3 1.25e+3   395.    14.4  1734.    1441.
#> 5       2 023033…     3.53 1492.  3.82e+3 1.25e+3   395.    14.4  1734.    1441.
#> 6       1 LTARPON    43.2    12.3 9.32e+2 5.29e+2   163.     2.50    9.96   609.
#> # ℹ 180 more variables: C_C02C <dbl>, C_C02D <dbl>, `C_C03 ` <dbl>,
#> #   C_C03A <dbl>, C_C03B <dbl>, C_C03C <dbl>, C_C03D <dbl>, `C_C04 ` <dbl>,
#> #   C_C04A <dbl>, C_C04B <dbl>, C_C04C <dbl>, C_C04D <dbl>, `C_C05 ` <dbl>,
#> #   C_C05A <dbl>, C_C05B <dbl>, C_C05C <dbl>, C_C05D <dbl>, `C_C06 ` <dbl>,
#> #   C_C06A <dbl>, C_C06B <dbl>, C_C06C <dbl>, C_C06D <dbl>, `C_C07 ` <dbl>,
#> #   C_C07A <dbl>, C_C07B <dbl>, C_C07C <dbl>, C_C07D <dbl>, `C_C08 ` <dbl>,
#> #   C_C08A <dbl>, C_C08B <dbl>, C_C08C <dbl>, C_C08D <dbl>, `C_C09 ` <dbl>, …